llm_service/requirements.txt
# Core FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# HTTP client libraries
aiohttp>=3.9.0
httpx>=0.25.0
requests>=2.31.0

# AI and ML dependencies
# Pinecone for vector database (v3+ API)
pinecone>=3.0.0

# Optional: If you want to use LangChain later
langchain>=0.1.0
langchain-huggingface>=0.0.3

# Optional: OpenAI client (if you want to switch to OpenAI embeddings)
openai>=1.0.0

# Database dependencies (if needed for other features)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

numpy
pydantic_settings
langchain_google_genai